+++ js/packages/react/src/index.ts
@@ -1,3 +1,5 @@
+'use client'
+
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2026 Barkpark contributors